+++ packages/create-sdk/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @tailor-platform/create-sdk
 
+## 1.7.0
+
 ## 1.6.3
 
 ## 1.6.2